fix(backend): resolve circular dependencies and duplicate logic in sockets
- Extract room state management into roomStateManager.js
- Update all socket handlers to import state from roomStateManager
- Remove redundant clear canvas validation check in whiteboardHandler
